pyproject.toml
[tool.poetry]
name = "AILiteSubtitler"
version = "0.0.5"
description = "AI千言字幕"
authors = ["Y00UU <12345@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydub = "^0.25.1"
openai = "^1.65.3"
retry = "^0.9.2"
requests = "^2.32.3"
pillow = "^11.1.0"
pysrt = "^1.1.2"
pyqt6-tools = "^6.4.2.3.3"

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
